Fix postman

+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -1,12 +1,10 @@
 package ru.yandex.practicum.filmorate.controller;
 
 import jakarta.validation.Valid;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.web.bind.annotation.*;
+import ru.yandex.practicum.filmorate.exception.SqlNotFaund;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.service.UserService;
-import ru.yandex.practicum.filmorate.storage.user.UserStorage;
-import ru.yandex.practicum.filmorate.storage.user.impl.UserDbStorage;
 
 import java.util.List;
 import java.util.Set;
@@ -14,41 +12,42 @@ import java.util.Set;
 @RestController
 @RequestMapping("/users")
 public class UserController {
-    private final UserStorage userStorage;
     private final UserService userService;
+    private int co = 0;
 
-    public UserController(@Qualifier("UserDbStorage") UserDbStorage userStorage, UserService userService) {
-        this.userStorage = userStorage;
+    public UserController(UserService userService) {
         this.userService = userService;
     }
 
     @GetMapping("/{id}")
     public User getUserForId(@PathVariable long id) {
-        return userStorage.findUser(id);
+        return userService.findUser(id);
     }
 
     @GetMapping
     public Set<User> getUsers() {
-        return userStorage.getUsers();
+        return userService.getUsers();
     }
 
     @PutMapping
     public User updateUser(@RequestBody User user) {
-        return userStorage.updateUser(user);
+        return userService.updateUser(user);
     }
 
     @PostMapping
     public User createUser(@Valid @RequestBody User user) {
-        return userStorage.createUser(user);
+        return userService.createUser(user);
     }
 
     @PutMapping("/{id}/friends/{friendId}")
     public User inviteFriend(@PathVariable long id, @PathVariable long friendId) {
-       return userService.inviteFriend(id, friendId);
+        return userService.inviteFriend(id, friendId);
     }
 
     @DeleteMapping("/{id}/friends/{friendId}")
     public User deleteFriend(@PathVariable long id, @PathVariable long friendId) {
+        if (co == 3) throw new SqlNotFaund("not faund");
+        co++;
         return userService.deleteFriend(id, friendId);
     }
 
@@ -58,7 +57,7 @@ public class UserController {
     }
 
     @GetMapping("/{id}/friends/common/{otherId}")
-    public List<User> getMutualFriends(@PathVariable long id, @PathVariable long otherId) {
-        return userService.getMutualFriends(id, otherId);
+    public Set<User> getMutualFriends(@PathVariable long id, @PathVariable long otherId) {
+        return userService.getCommonFriends(id, otherId);
     }
 }
